Prefault keypaths to avoid excessive KVO on faults at runtime
Ensure that all objects are faulted in (in the case of Core Data objects)
at init time, so that we don't have faults firing at unexpected times
at runtime. Note that we only fault in child key paths explicitly - sort
and filter key paths are already accessed as a matter of course.